// frontend/src/backend.d.ts
export interface HauntedHouseCharacter {
    voiceType: VoiceType;
    name: string;
    description: string;
    scareType: ScareType;
}
export type ContentType = {
    __kind__: "hauntedHouse";
    hauntedHouse: HauntedHouseSpecificFields;
} | {
    __kind__: "show";
    show: ShowSpecificFields;
} | {
    __kind__: "event";
    event: EventSpecificFields;
} | {
    __kind__: "scareZone";
    scareZone: ScareZoneSpecificFields;
} | {
    __kind__: "attraction";
    attraction: AttractionSpecificFields;
};
// ...
export interface ContentItem {
    id: bigint;
    useMainHauntSchedule: boolean;
    name: string;
    description: string;
    customType: ContentType;
    dates: Array<EventDateRange>;
}
// ...
export interface EventSpecificFields {
    eventType: EventType;
}
// ...
export interface ShowSpecificFields {
    yearIntroduced?: bigint;
    performanceType: PerformanceType;
}
export interface HauntedHouseSupervisorAuditionForm {
    age: bigint;
    name: string;
    email: string;
    experience: string;
    availability: string;
    phone: string;
    leadershipExperience: string;
}
export interface EventUnlockStatus {
    hasSecretEntrance: boolean;
    hasFireworks: boolean;
    hasFlynAppearance: boolean;
}
export interface Date_ {
    day: bigint;
    month: bigint;
    year: bigint;
}
// ...
export interface ScareZoneSpecificFields {
    yearIntroduced?: bigint;
    indoorOutdoor: ZoneLocation;
    scareLevel: ScareLevel;
}
export interface EventDateRange {
    endDate: Date_;
    startDate: Date_;
}
export interface AttractionSpecificFields {
    yearIntroduced?: bigint;
    ageRestriction: AgeRestriction;
    hasGuidedTour: boolean;
}
export interface HauntedHouseSpecificFields {
    yearIntroduced?: bigint;
    tagline: string;
    characters: Array<HauntedHouseCharacter>;
    sceneDescriptions: Array<string>;
    scareLevel: ScareLevel;
}
export interface Product {
    name: string;
    description: string;
    availability: AvailabilityStatus;
    category: ProductCategory;
    price: number;
}
export interface UserProfile {
    name: string;
}
export enum AgeRestriction {
    teens = "teens",
    kids = "kids",
    none = "none",
    adultsOnly = "adultsOnly"
}
export enum AuditionType {
    scareActor = "scareActor",
    costumeCharacter = "costumeCharacter",
    usher = "usher",
    danceActor = "danceActor",
    hauntedHouseSupervisor = "hauntedHouseSupervisor"
}
export enum AvailabilityStatus {
    inStock = "inStock",
    outOfStock = "outOfStock",
    limited = "limited"
}
export enum EventType {
    seasonal = "seasonal",
    convention = "convention",
    holiday = "holiday",
    specialEvent = "specialEvent",
    special = "special"
}
export enum FoodTheme {
    paint = "paint",
    death = "death"
}
export enum FoodType {
    dessert = "dessert",
    food = "food",
    drink = "drink"
}
export enum LandStatus {
    open = "open",
    underConstruction = "underConstruction",
    planned = "planned"
}
export enum PerformanceType {
    interactive = "interactive",
    theatrical = "theatrical",
    stunt = "stunt",
    dance = "dance",
    musical = "musical"
}
export enum ProductCategory {
    posters = "posters",
    skateboard = "skateboard",
    clothes = "clothes",
    shoes = "shoes",
    autograph = "autograph"
}
export enum ScareLevel {
    mild = "mild",
    extreme = "extreme",
    moderate = "moderate"
}
// ...
export enum ZoneLocation {
    both = "both",
    indoor = "indoor",
    outdoor = "outdoor"
}